refactor(navbar): user bubble component
