+++ components/navbar/UserBubble.js
@@ -1,6 +1,29 @@
+import * as React from "react";
+import axios from "axios";
 import {Box, IconButton, Typography, Menu, Avatar, Button, Tooltip, MenuItem} from "@mui/material";
+import { useOperation, useOperationMethod } from "react-openapi-client";
 
-function UserBubble({sessionActions}) {
+function UserBubble({sessionActions, userName, userId}) {
+  const [anchorElUser, setAnchorElUser] = React.useState(null);
+  const [endSession] = useOperationMethod("endSession");
+  
+  const handleOpenUserMenu = (event) => {
+    setAnchorElUser(event.currentTarget);
+  };
+  
+  const handleCloseUserMenu = () => {
+    setAnchorElUser(null);
+  };
+
+  const HandleSignOut = async () => {
+    await endSession();
+    if (window.location.pathname === "/") {
+      window.location.reload();
+    } else {
+      axios.get("/");
+    }
+  };
+  
   return(
     <Box sx={{ flexGrow: 0 }}>
       <Tooltip title="Manage session">
